Refactor to extract functions to facilitate their re-use.
Extract the functions dataset, snapshot and snapshot_rollback to
facilitate their re-use.

(cherry picked from commit 6e9868ed6dcd2cd2f809c0af06983524e0fdadfe)
